renamed dash details modal for vehicles
